requirements.txt
python-crontab==2.5.1
smartcitizen-connector
scdata
paho-mqtt
gunicorn
flask
invoke
celery
flower

staplus-api-python-client@git+https://github.com/securedimensions/STAplus-Python-Client@STAplus
oauth2-client@git+https://github.com/securedimensions/OAuth2Client.git
cryptography==3.4.6
PyJWT
hookdns